safe_conversions: use std::from_chars instead of std::stoi
std::stoi, std::stol, std::stoul, and std::stoull throw on malformed or
out-of-range input. When the input is sourced from an HTTP path/body or
a D-Bus property and the call sits inside an async callback, the
exception propagates uncaught and aborts bmcweb (SIGABRT). systemd
recovers via restart, but every active session is dropped and rapid
repeats can exhaust the restart policy.

The fix does not wrap std::stoi in a try/catch — it replaces the call
entirely with std::from_chars, which reports failure via std::errc and
never throws. Extend redfish::safe_conversions with:
  - stringToInt<T>(sv, base = 10): std::from_chars wrapper returning
    std::optional<T>; std::nullopt if the input is empty, has trailing
    garbage, or overflows T.
  - stringToHex<T>(sv): convenience wrapper that strips an optional
    "0x"/"0X" prefix and calls stringToInt<T>(sv, 16).

Migrate every std::stoi/stol/stoll/stoul/stoull call site in the tree
(52 sites across 20 files) to the helper. The accompanying try/catch
scaffolding is removed because the helper cannot throw; callers branch
on the optional instead. After the change there are no remaining
std::sto<int> calls outside the helper's own documentation comment.

This addresses Pattern 1 of bug 5914297 (HTTP-input std::stoi/stol
crashes) and the D-Bus AdditionalData std::stoi case (EVENT_DIR,
GENERATOR_ID) listed under Pattern 3. The remaining
nlohmann::json::items() / json::parse / json::dump unhandled-exception
sites from the same bug will be addressed in follow-up commits.

Tested:
  - bmcweb builds clean.
  - GET /redfish/v1/Managers/<m>/DedicatedNetworkPorts/NotANumber now
    returns a Redfish error (resourceNotFound) instead of crashing the
    service; bmcweb stays up; no SIGABRT in journal.
  - GET /redfish/v1/Systems/<s>/PCIeDevices/<d>/PCIeFunctions/abc and
    GET .../DebugTokenService/Entries/abc behave the same way.
  - Existing positive paths (numeric IDs, valid hex nonces in
    UpdateMinimumSecurityVersion / RevokeKeys) unchanged.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5914297

+++ include/debug_token/vdm/install_utils.hpp
@@ -34,7 +34,7 @@ enum VdmDebugTokenInstallErrorCode
     InternalError = 7
 };
 
-inline std::string getVdmDebugTokenInstallErrorDescription(int code)
+inline std::string getVdmDebugTokenInstallErrorDescription(int64_t code)
 {
     static std::map<VdmDebugTokenInstallErrorCode, std::string>
         debugTokenInstallErrorMap{
